% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcfPlot.R
\name{AcfPlot}
\alias{AcfPlot}
\title{Basic ACF Plotting}
\usage{
AcfPlot(g, LagZeroQ= TRUE, ylab=NULL, main=NULL, ...)
}
\arguments{
\item{g}{vector of autocorrelations at lags 1,..,length(g).}

\item{LagZeroQ}{start plot at lag zero with g[0]=1.}

\item{ylab}{vertical axis label.}

\item{main}{plot title.}

\item{...}{optional graphical parameters.}
}
\value{
No value. Plot is produced via plot function.
}
\description{
Produces theoretical correlation plot
}
\examples{
# Simple example, plot acf for AR(1)
phi<-0.8
maxLag<-20
g<-phi^(1:maxLag)
AcfPlot(g)
AcfPlot(g, LagZeroQ=FALSE)

Plot the sample inverse partial autocorrelations
On the basis of this plot, Cleveland (1972) suggested an ARp(1,2,7)
for this data
"InverseAcf" <-
function(z, p=15){
    g<-TacvfMA(GetFitARpLS(z-mean(z),1:p)$phiHat, lag.max=p)
    g/g[1]
}
data(SeriesA)
AcfPlot(InverseAcf(SeriesA),LagZeroQ=FALSE)

}
\seealso{
\code{\link[=acf]{acf()}}.
}
\author{
A.I. McLeod and Y. Zhang.
}
