% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARSdf.R
\name{ARSdf}
\alias{ARSdf}
\title{Autoregressive Spectral Density Function}
\usage{
ARSdf(phi, pFFT = 8)
}
\arguments{
\item{phi}{AR Coefficient vector.}

\item{pFFT}{FFT with $2^pFFF$ FFF frequencies, default 8.}
}
\value{
A vector of the density function values, \eqn{(f(1),\dots,f(2^p FFF))}.
}
\description{
Spectral density function of AR(p) is computed.
}
\details{
The Fast Fourier Transform (FFT) is used to compute the spectral density function.
}
\examples{
ARSdf(0.8)
ARSdf(c(0.1,0.2))

}
\seealso{
\code{\link[=spectrum]{spectrum()}}, \code{\link[=spec.pgram]{spec.pgram()}}, \code{\link[=spec.ar]{spec.ar()}}.
}
\author{
A.I. McLeod and Y. Zhang.
}
