% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARToMA.R
\name{ARToMA}
\alias{ARToMA}
\title{Coefficients In Infinite Moving Average Expansion}
\usage{
ARToMA(phi, lag.max)
}
\arguments{
\item{phi}{AR Coefficient vector.}

\item{lag.max}{maximum lag.}
}
\value{
Vector of length L+1 containing, \eqn{(1,\psi_1,\dots,\psi_L)}.
}
\description{
A stationary-causal AR(p) can be written as a general linear process (GLP).
This function obtains the moving-average expansion out to the L-th lag,
\eqn{z_t=a_t+\psi_1 a_{t-1}+\cdots+\psi_L a_{t-L}}.
}
\details{
The coefficients are computed recursively as indicated in Box and Jenkins (1970).
}
\examples{
ARToMA(0.5,20)
ARToMA(c(0.2,0.5), 15)

}
\references{
Box and Jenkins (1970), Time Series Analysis, Forecasting & Control.
}
\seealso{
\code{\link[=InvertibleQ]{InvertibleQ()}}.
}
\author{
A.I. McLeod and Y. Zhang.
}
