% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARToPacf.R
\name{ARToPacf}
\alias{ARToPacf}
\title{Reparametrize AR Coefficients In Terms of PACF}
\usage{
ARToPacf(phi)
}
\arguments{
\item{phi}{vector of AR parameter coefficients.}
}
\value{
Vector of length(phi) containing the parameters in the transformed PACF domain.
}
\description{
Transform AR parameter coefficients into partial autocorrelation function (PACF).

No check for invertibility is done for maximum computational efficiency
since this function is used extensively in the numerical optimization
of the AR loglikelihood function in FitAR.
Use InvertibleQ to test for invertible AR coefficients.
}
\details{
For details see McLeod and Zhang (2006).
}
\examples{
somePACF<-c(0.5,0.6,0.7,0.8,-0.9,-0.8)
# PacfToAR() transforms PACF to AR parameter coefficients.
someAR<-PacfToAR(somePACF)
test<-ARToPacf(someAR)
#This should be very small
sum(abs(test-somePACF))

}
\references{
McLeod, A.I. and Zhang, Y. (2006). Partial autocorrelation parameterization
for subset autoregression. Journal of Time Series Analysis, 27, 599-612.
}
\seealso{
\code{\link[=InvertibleQ]{InvertibleQ()}}, \code{\link[=PacfToAR]{PacfToAR()}}.
}
\author{
A.I. McLeod and Y. Zhang.
}
