% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackcastResidualsAR.R
\name{BackcastResidualsAR}
\alias{BackcastResidualsAR}
\title{Innovation Residuals in AR}
\usage{
BackcastResidualsAR(y, phi, Q = 100, demean=TRUE)
}
\arguments{
\item{y}{a time series or vector.}

\item{phi}{AR coefficients, lags 1,...,p.}

\item{Q}{for backcasting, the AR is approximated by an MA(Q).}

\item{demean}{subtract sample mean.}
}
\value{
Vector of residuals.
}
\description{
Obtains the residuals (estimated innovations).
The residuals for t=1,...,p are obtained using
the backforecasting algorithm of Box and Jenkins (1970).
}
\details{
The backforecasting algorithm is described in detail in the book of Box and
Jenkins (1970). The idea is to compute the expected value of the innovation
assuming a high-order MA(q).
}
\note{
No check is done that the AR is causal-stationary.
}
\examples{
# compare residuals obtained using backcasting with fitted parameters and
# the residuals extracted from output of FitAR. They are identical.
p<-11
out<-FitAR(log(lynx), p)
phi<-out$phiHat # fitted parameters
resphi<-BackcastResidualsAR(log(lynx), phi)
sum(abs(resphi-resid(out)))

}
\references{
Box and Jenkins (1970). Time Series Analysis: Forecasting and Control.
}
\seealso{
\code{\link[=InvertibleQ]{InvertibleQ()}}, \code{\link[=FitAR]{FitAR()}}.
}
\author{
A.I. McLeod and Y. Zhang.
}
