% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BICqSelect.R
\name{BICqSelect}
\alias{BICqSelect}
\title{Select best model using BICq}
\usage{
BICqLL(logL, n, level = 0.99, mSize = 1:length(logL), mComplex = function(k) k)
}
\arguments{
\item{logL}{vector of loglikelihoods.}

\item{n}{sample size.}

\item{level}{probability.}

\item{mSize}{model sizes.}

\item{mComplex}{a complexity function.}
}
\value{
\itemize{
\item \code{khat} dataframe with columns: k, a.1, a.2 q.1, q.2, level=level, where k is the
optimal model, (a.1,a.2) is the interval for alpha in the GIC, (q.1, q.2) is
the interval for q and level is the probability. Each row corresponds to
an entry in 'level'.
\item \code{table} This table indicates which models can be selected for some values of
alpha or q.
}
}
\description{
Given the loglikelihoods for a set of models arranged in ascending order of
size, the best model is selected using the BICq criterion for a specified size.
}
\details{
See reference.
}
\note{
AIC corresponds to setting level=0.84. BIC corresponds to setting
level=pchisq(log(n), 1). So for n=100, 1000; BIC=0.96, 0.97.
}
\examples{
# Example 1.
# AR(p) Order selection for 'lynx' series
z <- log(lynx)
n <- length(z)
lag.max <- 20
zta<-ARToPacf(ar.burg(z,aic=FALSE,order.max=lag.max)$ar)
LagsEntering<-1:lag.max
LLapprox <- (-n)*log(cumprod(1-zta[LagsEntering]^2))
ans<-BICqLL(logL=LLapprox, n=n, level=c(0.9, 0.95, 0.99))
ans$khat
ans$table
# if we just want the best model for level=0.99 then,
(BICqSelect(logL=LLapprox, n=n, level=0.99)$khat)[[1]]
# aic for comparison
aic<-(-2*LLapprox)+2*LagsEntering
which.min(aic)
plot(LagsEntering, aic)

}
\references{
Changjiang Xu and A. Ian McLeod (2010). Bayesian information criterion
Bernoulli prior. Submitted for publication.
}
\seealso{
\code{\link[=SelectModel]{SelectModel()}}.
}
\author{
Changjiang Xu and A. Ian McLeod.
}
