% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boot.R
\name{Boot}
\alias{Boot}
\title{Generic Bootstrap Function}
\usage{
Boot(obj, R=1, ...)
}
\arguments{
\item{obj}{fitted object.}

\item{R}{number of bootstrap replications.}

\item{...}{optional arguments.}
}
\value{
Parametric bootstrap simulation.
}
\description{
Generic function to bootstrap a fitted model.
}
\details{
At present, the only function implemented is \code{Boot.FitAR}.
}
\examples{
out<-FitAR(SeriesA, c(1,2,7), ARModel="ARp")
Boot(out)

}
\seealso{
\code{\link[=Boot.FitAR]{Boot.FitAR()}}.
}
\author{
A.I. McLeod and Y. Zhang.
}
