% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boot.ts.R
\name{Boot.ts}
\alias{Boot.ts}
\title{Parametric Time Series Bootstrap}
\usage{
Boot(obj, R=1, ...) # S3 method for class 'ts'
}
\arguments{
\item{obj}{a time series, class "ts".}

\item{R}{number of bootstrap replications.}

\item{...}{optional arguments.}
}
\value{
A time series or vector.
}
\description{
An AR(p) model is fit to the time series using the AIC and then it is simulated.
}
\note{
Parametric and nonparametric time series bootstraps are discussed
by Davison and Hinkley (1997, Ch.8.2). Nonparametric bootstrap
for time series is available in the function tsboot in the library boot.
}
\examples{
layout(matrix(c(1,2,1,2),ncol=2))
TimeSeriesPlot(SeriesA)
TimeSeriesPlot(Boot(SeriesA),main="Bootstrap of Series A")

}
\references{
Davison, A.C. and Hinkley, D.V. (1997), Bootstrap Methods
and Their Application. Cambridge University Press.
}
\seealso{
\code{\link[=Boot.FitAR]{Boot.FitAR()}}.
}
\author{
A.I. McLeod and Y. Zhang.
}
