% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxCox.FitAR.R
\name{BoxCox.FitAR}
\alias{BoxCox.FitAR}
\title{Box-Cox Analysis for "FitAR" Objects}
\usage{
BoxCox(object, interval = c(-1, 1), type = "BoxCox", InitLambda = "none", ...) ## S3 method for class 'FitAR'
}
\arguments{
\item{object}{output from FitAR.}

\item{interval}{interval to be searched for the optimal transformation.}

\item{type}{ignored unless, InitLambda!="none". Type of transformation,
default is "BoxCox". Otherwise a simple power transformation.}

\item{InitLambda}{default "none". Otherwise a numerical value
giving the transformation parameter.}

\item{...}{optional arguments passed to optimize.}
}
\value{
No value returned. Graphical output is produced as side-effect.
The plot shows relative likelihood function as well as the MLE and a confidence interval.
}
\description{
This is a methods function to do a Box-Cox analysis for models fit using FitAR.
}
\details{
If no transformation is used on the data, then the original data is used.
But if a transformation has already been used, we need to inverse
transform the data to recover the untransformed data.

For \eqn{\lambda \ne 0}, the Box-Cox transformation is of x is \eqn{(x^{\lambda}-1)/\lambda)}.
If the minimum data value is <= 0, a small positive constant, equal
to the negative of the minimum plus 0.25, is added to all the data values.
}
\note{
The MASS package has a similar function \code{boxcox}
but this is implemented only for regression and analysis of variance.
}
\examples{
## Not run:  # takes a few seconds
# lynx time series. ARp subset model.
out<-FitAR(lynx, c(1,2,4,10,11), ARModel="ARp")
BoxCox(out)
#
p<-SelectModel(lynx, ARModel="ARz", lag.max=25, Best=1)
out<-FitAR(lynx, p)
BoxCox(out)
## End(Not run)

}
\references{
Box, G. E. P. and Cox, D. R. (1964) An analysis of transformations.
Journal of Royal Statistical Society, Series B, vol. 26, pp. 211-246.
}
\seealso{
\code{\link[=BoxCox]{BoxCox()}}, \code{\link[=BoxCox.Arima]{BoxCox.Arima()}}.
}
\author{
A.I. McLeod.
}
