% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxCox.R
\name{BoxCox}
\alias{BoxCox}
\title{Generic Box-Cox Analysis Function}
\usage{
BoxCox(object, ...)
}
\arguments{
\item{object}{model object.}

\item{...}{optional arguments.}
}
\value{
No value returned. Graphical output is produced as side-effect.
The plot shows relative likelihood function as well as the MLE and a confidence interval.
}
\description{
The function is implemented as a generic function with methods for classes
"FitAR", "Arima", "ts" and "numeric".
}
\details{
For \eqn{\lambda \ne 0}, the Box-Cox transformation is of x is \eqn{(x^{\lambda}-1)/\lambda)}.
If the minimum data value is <= 0, a small positive constant, equal
to the negative of the minimum plus 0.25, is added to all the data values.
}
\note{
The MASS package has a similar function \code{boxcox}
but this is implemented only for regression and analysis of variance.
}
\examples{
## Not run:  # takes a few seconds
BoxCox(lynx)
out<-FitAR(lynx, c(1,2,4,10,11), ARModel="ARp", MLEQ=FALSE)
BoxCox(out)
out<-FitAR(lynx, c(1,2,4,5,7,8,10,11,12))
BoxCox(out)
## End(Not run)

}
\seealso{
\code{\link[=BoxCox.Arima]{BoxCox.Arima()}}, \code{\link[=BoxCox.FitAR]{BoxCox.FitAR()}}, \code{\link[=BoxCox.ts]{BoxCox.ts()}}, \code{\link[=BoxCox.numeric]{BoxCox.numeric()}}.
}
\author{
A.I. McLeod and Y. Zhang.
}
