% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxCox.ts.R
\name{BoxCox.ts}
\alias{BoxCox.ts}
\title{Box-Cox Analysis for a Time Series}
\usage{
BoxCox(object, interval = c(-1, 1), ...) ## S3 method for class 'ts'
}
\arguments{
\item{object}{a vector of time series values.}

\item{interval}{interval to be searched.}

\item{...}{optional arguments.}
}
\value{
No value returned. Graphical output is produced as side-effect.
The plot shows relative likelihood function as well as the MLE and a confidence interval.
}
\description{
The time series is converted to a vector and BoxCox.numeric is used.
}
\details{
For \eqn{\lambda \ne 0}, the Box-Cox transformation is of x is \eqn{(x^{\lambda}-1)/\lambda)}.
If the minimum data value is <= 0, a small positive constant, equal
to the negative of the minimum plus 0.25, is added to all the data values.

It is important not to transform the data when fitting it with AR since the
optimal transformation would be found for the transformed data – not the
original data. Normally this would not be a sensible thing to do.
}
\note{
The MASS package has a similar function \code{boxcox}
but this is implemented only for regression and analysis of variance.
}
\examples{
## Not run:  # takes a few seconds
BoxCox(sunspot.year)
## End(Not run)

}
\references{
Box, G. E. P. and Cox, D. R. (1964) An analysis of transformations.
Journal of Royal Statistical Society, Series B, vol. 26, pp. 211-246.
}
\seealso{
\code{\link[=BoxCox.FitAR]{BoxCox.FitAR()}}, \code{\link[=BoxCox.Arima]{BoxCox.Arima()}}, \code{\link[=BoxCox.numeric]{BoxCox.numeric()}}.
}
\author{
A.I. McLeod
}
