% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.FitAR.R
\name{coef.FitAR}
\alias{coef.FitAR}
\title{Display Estimated Parameters from Output of "FitAR"}
\usage{
coef(object, ...) ## S3 method for class 'FitAR'
}
\arguments{
\item{object}{obj the output from FitAR.}

\item{...}{optional parameters.}
}
\value{
A matrix is returned. The columns of the matrix are labeled MLE, sd and Z-ratio.
The rows labels indicate the AR coefficients which were estimated followed by
mu, the estimate of mean.
}
\description{
Method function to display fitted parameters, their standard errors and
Z-ratio for AR models fit with FitAR.
}
\examples{
# Fit subset AR to SeriesA
outA<-FitAR(SeriesA, c(1,2,7), ARModel="ARz")
coef(outA)
#
outALS<-FitAR(SeriesA, c(1,2,7), ARModel="ARp")
coef(outALS)

}
\references{
McLeod, A.I. and Zhang, Y. (2006). Partial autocorrelation parameterization
for subset autoregression. Journal of Time Series Analysis, 27, 599-612.
}
\author{
A.I. McLeod and Y. Zhang.
}
