% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitARp.R
\name{FitARp}
\alias{FitARp}
\title{Fit subset ARp Models}
\usage{
FitARp(z, p, lag.max = "default", MLEQ = FALSE)
}
\arguments{
\item{z}{time series, vector or ts object.}

\item{p}{p specifies the model. If length(p) is 1, an AR(p) is assumed and
if p has length greater than 1, a subset ARp is assumed. For example,
to fit a subset model with lags 1 and 4 present, set p to c(1,4) or
equivalently c(1,0,0,4). To fit a subset model with just lag 4,
you must use p=c(0,0,0,4) since p=4 will fit a full AR(4).}

\item{lag.max}{the residual autocorrelations are tabulated for lags
1, ..., lag.max. Also lag.max is used for the Ljung-Box portmanteau test.}

\item{MLEQ}{TRUE, use MLE. FALSE, use LS.}
}
\value{
A list with class name "FitAR" and components:
\itemize{
\item \code{loglikelihood} value of the loglikelihood.
\item \code{phiHat} coefficients in AR(p) – including 0's.
\item \code{sigsqHat} innovation variance estimate.
\item \code{muHat} estimate of the mean.
\item \code{covHat} covariance matrix of the coefficient estimates.
\item \code{zetaHat} transformed parameters, length(zetaHat) = \# coefficients estimated.
\item \code{RacfMatrix} residual autocorrelations and sd for lags 1, ..., lag.max.
\item \code{LjungBox} table of Ljung-Box portmanteau test statistics.
\item \code{SubsetQ} parameters in AR(p) – including 0's.
\item \code{res} innovation residuals, same length as z.
\item \code{fits} fitted values, same length as z.
\item \code{pvec} lags used in AR model.
\item \code{demean} TRUE if mean estimated otherwise assumed zero.
\item \code{FitMethod} "MLE" or "LS".
\item \code{IterationCount} number of iterations in mean mle estimation.
\item \code{convergence} value returned by optim – should be 0.
\item \code{MLEMeanQ} TRUE if mle for mean algorithm used.
\item \code{ARModel} "ARp" if FitARp used, otherwise "ARz".
\item \code{tsp} tsp(z).
\item \code{call} result from match.call() showing how the function was called.
\item \code{ModelTitle} description of model.
\item \code{DataTitle} returns attr(z,"title").
\item \code{z} time series data input.
}
}
\description{
The subset ARp is defined as an AR(p) in which some of the ar-coefficients
are constrained to zero. This is the usual type of subset AR. In contrast
the ARz model constrains some of the partial autocorrelation coefficients to zero.
}
\details{
Subset ARp model is fit using exact MLE. The built-in arima function is used
for MLE. When MLEQ=FALSE, LS is used. LS is has been widely used in past for
subset ARp fiting.
}
\examples{
# First example: Fit to AR(4) 
set.seed(3323)
phi<-c(2.7607,-3.8106,2.6535,-0.9238)
z<-SimulateGaussianAR(phi,1000)
# MLE using arima
ans1<-FitARp(z,4,MLEQ=TRUE)
ans1
coef(ans1)
# OLS
ans2<-FitARp(z,4,MLEQ=FALSE)
ans2
coef(ans2)

## Not run: # save time building package
# Second Example: Fit subset ARp model
z<-log(lynx)
# MLE 
FitARp(z, c(1,2,4,7,10,11),MLEQ=TRUE)
# LS
FitARp(z, c(1,2,4,7,10,11),MLEQ=FALSE)

# Third Example: Use UBIC model selection to fit subset models
z<-log(lynx)
p<-SelectModel(z,ARModel="ARp")[[1]]$p
# MLE # error returned by arima
# ans1<-FitARp(z, p, MLEQ=TRUE)
# ans1
# LS
ans2<-FitARp(z, p, MLEQ=FALSE)
ans2

}
\references{
McLeod, A.I. and Zhang, Y. (2006). Partial autocorrelation
parameterization for subset autoregression. Journal of Time Series Analysis, 27, 599-612.
}
\seealso{
\code{\link[=FitAR]{FitAR()}}, \code{\link[=FitARz]{FitARz()}}, \code{\link[=GetFitARz]{GetFitARz()}}, \code{\link[=GetFitARpMLE]{GetFitARpMLE()}}, \code{\link[=RacfPlot]{RacfPlot()}}.
}
\author{
A.I. McLeod.
}
