% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitARz.R
\name{FitARz}
\alias{FitARz}
\title{Subset ARz Model Fitting}
\usage{
FitARz(z, p, demean = TRUE, MeanMLEQ = FALSE, lag.max = "default")
}
\arguments{
\item{z}{time series, vector or ts object.}

\item{p}{p specifies the model. If length(p) is 1, an AR(p) is assumed and
if p has length greater than 1, a subset ARz is assumed. For example,
to fit a subset model with lags 1 and 4 present, set p to c(1,4) or
equivalently c(1,0,0,4). To fit a subset model with just lag 4,
you must use p=c(0,0,0,4) since p=4 will fit a full AR(4).}

\item{demean}{TRUE, mean estimated. FALSE, mean is zero.}

\item{MeanMLEQ}{use exact MLE for mean parameter.}

\item{lag.max}{the residual autocorrelations are tabulated for lags
1, ..., lag.max. Also lag.max is used for the Ljung-Box portmanteau test.}
}
\value{
A list with class name "FitAR" and components:
\itemize{
\item \code{loglikelihood} value of the loglikelihood.
\item \code{phiHat} coefficients in AR(p) – including 0's.
\item \code{sigsqHat} innovation variance estimate.
\item \code{muHat} estimate of the mean.
\item \code{covHat} covariance matrix of the coefficient estimates.
\item \code{zetaHat} transformed parameters, length(zetaHat) = \# coefficients estimated.
\item \code{RacfMatrix} residual autocorrelations and sd for lags 1, ..., lag.max.
\item \code{LjungBox} table of Ljung-Box portmanteau test statistics.
\item \code{SubsetQ} parameters in AR(p) – including 0's.
\item \code{res} innovation residuals, same length as z.
\item \code{fits} fitted values, same length as z.
\item \code{pvec} lags used in AR model.
\item \code{demean} TRUE if mean estimated otherwise assumed zero.
\item \code{FitMethod} "MLE" or "LS".
\item \code{IterationCount} number of iterations in mean mle estimation.
\item \code{convergence} value returned by optim – should be 0.
\item \code{MLEMeanQ} TRUE if mle for mean algorithm used.
\item \code{ARModel} "ARp" if FitARp used, otherwise "ARz".
\item \code{tsp} tsp(z).
\item \code{call} result from match.call() showing how the function was called.
\item \code{ModelTitle} description of model.
\item \code{DataTitle} returns attr(z,"title").
\item \code{z} time series data input.
}
}
\description{
The subset ARz model, defined by constraining partial autocorrelations to
zero, is fitted using exact MLE. When length(p)=1, an AR(p) is fit by MLE.
}
\details{
The model and its properties are discussed in McLeod and Zhang (2006)
and McLeod and Zhang (2008).
}
\note{
Normally one would use the \code{FitAR} function which then calls this
function for the ARz case.
}
\examples{
# First example: Fit exact MLE to AR(4) 
set.seed(3323)
phi<-c(2.7607,-3.8106,2.6535,-0.9238)
z<-SimulateGaussianAR(phi,1000)
ans<-FitARz(z,4,MeanMLEQ=TRUE)
ans
coef(ans)

## Not run: # save time building package
# Second Example: compare with sample mean result
ans<-FitARz(z,4)
coef(ans)

# Third Example: Fit subset ARz
z<-log(lynx)
FitARz(z, c(1,2,4,7,10,11))
# now obain exact MLE for Mean as well
FitARz(z, c(1,2,4,7,10,11), MeanMLE=TRUE)

# Fourth Example: Fit subset ARz
somePACF<-c(0.5,0,0,0,-0.9)
someAR<-PacfToAR(somePACF)
z<-SimulateGaussianAR(someAR,1000)
ans=FitARz(z, c(1,5),MeanMLEQ=TRUE)
coef(ans)
GetFitARz(z,c(1,5))#assuming a known zero mean

## End(Not run)

}
\references{
McLeod, A.I. and Zhang, Y. (2006). Partial autocorrelation
parameterization for subset autoregression. Journal of Time Series Analysis, 27, 599-612.
}
\seealso{
\code{\link[=FitAR]{FitAR()}}, \code{\link[=FitARp]{FitARp()}}, \code{\link[=GetFitARz]{GetFitARz()}}, \code{\link[=GetFitARpMLE]{GetFitARpMLE()}}, \code{\link[=RacfPlot]{RacfPlot()}}.
}
\author{
A.I. McLeod.
}
