% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.FitAR.R
\name{fitted.FitAR}
\alias{fitted.FitAR}
\title{Fitted Values from "FitAR" Object}
\usage{
fitted(object, ...) ## S3 method for class 'FitAR'
}
\arguments{
\item{object}{object of class "FitAR".}

\item{...}{optional arguments.}
}
\value{
Vector of fitted values.
}
\description{
Method function, extracts fitted values from FitAR object.
}
\examples{
out<-FitAR(SeriesA, c(1,2,6,7))
fitted(out)

}
\references{
McLeod, A.I. and Zhang, Y. (2006). Partial autocorrelation
parameterization for subset autoregression. Journal of Time Series Analysis, 27, 599-612.
}
\seealso{
\code{\link[=FitAR]{FitAR()}}.
}
\author{
A.I. McLeod and Y. zhang.
}
