% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FromSymmetricStorageUpper.R
\name{FromSymmetricStorageUpper}
\alias{FromSymmetricStorageUpper}
\title{Converts a Matrix from Symmetric Storage Mode to Regular Format}
\usage{
FromSymmetricStorageUpper(x)
}
\arguments{
\item{x}{a vector which represents a matrix in upper triangular form.}
}
\value{
symmetric matrix.
}
\description{
Utility function.
}
\examples{
FromSymmetricStorageUpper(1:5)

}
\author{
A.I. McLeod.
}
