% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetB.R
\name{GetB}
\alias{GetB}
\title{Internal Utility Function}
\usage{
GetB(phi)
}
\arguments{
\item{phi}{a vector of AR coefficients.}
}
\description{
The user would not normally use this function. The function is needed for
exact mle for mean. Used in Get1G which is called from GetARMeanMLE.
}
