% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFitAR.R
\name{GetFitAR}
\alias{GetFitAR}
\title{MLE for AR, ARp and ARz}
\usage{
GetFitAR(z, p, ARModel = "ARz", ...)
}
\arguments{
\item{z}{time series.}

\item{p}{model order or subset lags.}

\item{ARModel}{either "ARp" or "ARz" corresponding to \code{GetFitARp} or \code{GetFitARz}.}

\item{...}{optional arguments which are passed to \code{GetFitARp} or \code{GetFitARz}.}
}
\value{
The returned values are as follow:
\itemize{
\item \code{loglikelihood} value of maximized loglikelihood.
\item \code{zetaHat} estimated zeta parameters.
\item \code{phiHat} estimated phi parameters.
\item \code{convergence} result from optim.
}
}
\description{
Obtains the exact MLE for AR(p) or subset AR models ARp or ARz.
This function is used by FitAR. One might prefer to use GetFitAR
for applications such as bootstrapping since it is faster than FitAR.
}
\details{
This is just a shell which simply invokes either GetFitARp or GetFitARz.
}
\examples{
# compare results from GetFitAR and FitAR
z<-log(lynx)
z<-z - mean(z)
GetFitAR(z, c(1,2,8))
out<-FitAR(log(lynx), c(1,2,8))
out
coef(out) 

}
\references{
McLeod, A.I. and Zhang, Y. (2006). Partial autocorrelation
parameterization for subset autoregression. Journal of Time Series Analysis, 27, 599-612.
}
\seealso{
\code{\link[=FitAR]{FitAR()}}.
}
\author{
A.I. McLeod.
}
