% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFitARpMLE.R
\name{GetFitARpMLE}
\alias{GetFitARpMLE}
\title{Exact MLE for subset ARp Models}
\usage{
GetFitARpMLE(z, pvec)
}
\arguments{
\item{z}{time series.}

\item{pvec}{lags included in AR model. If pvec = 0, white noise model assumed.}
}
\value{
a list with components:
\itemize{
\item \code{loglikelihood} the exact loglikelihood.
\item \code{phiHat} estimated AR parameters.
\item \code{constantTerm} constant term in the linear regression.
\item \code{pvec} lags of estimated AR coefficient.
\item \code{res} the least squares regression residuals.
\item \code{InvertibleQ} True, if the estimated parameters are in the AR admissible region.
}
}
\description{
Uses built-in function \code{arima} to fit subset ARp model, that is,
the subset model is formed by constraining some coefficients to zero.
}
\details{
Due to the optimization algorithms used by \code{arima}, this method
is not very reliable. The optimization may simply fail. Example 1 shows
it working but in Example 2 below it fails.
}
\examples{
# Example 1. MLE works
z<-log(lynx)
p<-c(1,2,4,7,10,11)
GetFitARpMLE(z, p)
#
# Example 2. MLE fails with error.
p<-c(1,2,9,12)
## Not run: GetFitARpMLE(z, p)

}
\references{
McLeod, A.I. and Zhang, Y. (2006). Partial autocorrelation
parameterization for subset autoregression. Journal of Time Series Analysis, 27, 599-612.
}
\seealso{
\code{\link[=FitAR]{FitAR()}}, \code{\link[=FitARz]{FitARz()}}, \code{\link[=GetFitARz]{GetFitARz()}}, \code{\link[=FitARp]{FitARp()}}, \code{\link[=RacfPlot]{RacfPlot()}}.
}
\author{
A.I. McLeod.
}
