% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFitARz.R
\name{GetFitARz}
\alias{GetFitARz}
\title{Exact MLE for AR(p) and Subset ARz – Short Version}
\usage{
GetFitARz(z, pvec, MeanValue=0, ...)
}
\arguments{
\item{z}{time series.}

\item{pvec}{lags included in AR model. If pvec = 0, white noise model assumed.}

\item{MeanValue}{by default it is assumed the mean of z is 0.}

\item{...}{optional arguments passed through to optim.}
}
\value{
a list with components:
\itemize{
\item \code{loglikelihood} value of maximized loglikelihood.
\item \code{zetaHat} estimated zeta parameters.
\item \code{phiHat} estimated phi parameters.
\item \code{convergence} result from optim.
\item \code{pvec} lags of estimated AR coefficient.
\item \code{algorithm} "BFGS" or "Nelder-Mead".
}
}
\description{
Obtain the exact MLE for AR(p) or subset ARz model. This function is used
by \code{FitAR} and \code{FitARz}. One might prefer to use GetFitARz for applications
such as bootstrapping since it is faster.
}
\details{
The built-in function optim is used to obtain the MLE estimates
for an AR or subset AR. First "BFGS" is tried. This usually works fine.
In the rare cases where convergence is not obtained, "Nelder-Mead" is used.
A warning message is given if this happens.
}
\examples{
# compare results from GetFitARz and FitAR
z<-log(lynx)
z<-z - mean(z)
GetFitARz(z, c(1,2,8))
out<-FitAR(log(lynx), c(1,2,8), ARModel="ARz")
out
coef(out)

}
\references{
McLeod, A.I. and Zhang, Y. (2006). Partial autocorrelation
parameterization for subset autoregression. Journal of Time Series Analysis, 27, 599-612.
}
\seealso{
\code{\link[=FitAR]{FitAR()}}, \code{\link[=FitARz]{FitARz()}}, \code{\link[=GetFitARpMLE]{GetFitARpMLE()}}, \code{\link[=RacfPlot]{RacfPlot()}}.
}
\author{
A.I. McLeod and Y. Zhang.
}
