% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLeapsAR.R
\name{GetLeapsAR}
\alias{GetLeapsAR}
\title{Select lags for Best Subset ARp Model}
\usage{
GetLeapsAR(z, lag.max = 15, Criterion = "UBIC", Best = 3, Candidates=5, t="default", ExactQ=FALSE)
}
\arguments{
\item{z}{ts object or vector containing time series.}

\item{lag.max}{maximum order of the AR.}

\item{Criterion}{default UBIC, other choices are "AIC", "BIC", "EBIC", "BICq", "GIC".}

\item{Best}{the number of based selected. Ignore with "GIC".}

\item{Candidates}{number of models initially selected using the approximate criterion.}

\item{t}{tuning parameter, EBIC, BICq, GIC.}

\item{ExactQ}{exhaustive numeration using exact likelihood. Still under under development. NOT AVAILABLE IN THIS VERSION.}
}
\value{
When 'Criterion' is one of UBIC, AIC, BIC, EBIC, BICq, a list with components:
\itemize{
\item \code{p} lags present in model.
\item \code{UBIC} approximate UBIC (Chen & Chen, 2007), if Criterion=="UBIC".
\item \code{AIC} approximate AIC (McLeod and Zhang, 2006a, eqn. 15), if Criterion=="AIC".
\item \code{BIC} approximate BIC (McLeod and Zhang, 2006a, eqn. 15), if Criterion=="BIC".
\item \code{EBIC} approximate EBIC (McLeod and Zhang, 2006a, eqn. 15), if Criterion=="EBIC".
\item \code{BICq} approximate BICq, if Criterion=="BICq".
\item \code{GIC} approximate GIC, if Criterion=="GIC".
}
}
\description{
The subset ARp model is the usual subset model, for example see Tong (1977).
This function is used by SelectModel for model identification for ARp models.
}
\details{
The R function leaps in the R package leaps is used to compute the subset
regression model with the smallest residual sum of squares containing 1, ...,
lag.max parameters. The mean is always included, so the only parameters
considered are the phi coefficients. After the best models containing 1, ...,
lag.max parameters are selected the models are individually refit to determine
the exact likelihood function for each selected model. Based on this likelihood
the UBIC/BIC/AIC is computed and then the best models are selected. The UBIC
criterion was developed by Chen and Chen (2007). The EBIC using a tuning
parameter, G, where 0 <= G <= 1. The BICq takes a tuning parameter, Q, where
0 < Q < 1. The GIC takes a tuning parameter, p, where 0<p<0.25.
}
\note{
AIC and BIC values produced are not comparable to AIC and BIC produced by
SelectModel for ARz models. However comparable AIC/BIC values are produced
when the selected models are fit by FitAR.

Requires leaps package. Since the least-squares is used, the number of
observations depends on 'lag.max'. Hence different subsets may be chosen
depending on the 'lag.max. See example below.
}
\examples{
# Example 1: Simple Example
# for the log(lynx) Tong (1977) selected an ARp(1,2,4,10,11)
# using the AIC and a subset selection algorithm. Our more exact
# approach shows that the ARp(1,2,3,4,10,11) has slightly lower
# AIC (using exact likelihood evaluation).
z<-log(lynx)
GetLeapsAR(z, lag.max=11)
GetLeapsAR(z, lag.max=11, Criterion="BIC")

# Example 2: Subset autoregression depends on lag.max!
# Because least-squares is used, P=lag.max observations are deleted.
# This phenomenon does not happen with "ARz" subset models
# ARp models depend on lag.max
GetLeapsAR(z, lag.max=15, Criterion="BIC")
GetLeapsAR(z, lag.max=20, Criterion="BIC")

# Example 3: Comparing GIC with BIC, AIC, UBIC and BICq
z <- log(lynx)
GetLeapsAR(z, lag.max=15, Criterion="BIC", Best=1)
GetLeapsAR(z, lag.max=15, Criterion="AIC", Best=1)
GetLeapsAR(z, lag.max=15, Criterion="UBIC", Best=1)
GetLeapsAR(z, lag.max=15, Criterion="BICq", Best=1, t=0.25)
GetLeapsAR(z, lag.max=15, Best=1, Criterion="GIC", t=0.01)
ans<-GetLeapsAR(z, lag.max=15, Best=3, Criterion="GIC", t=0.001)
plot(ans)

}
\references{
Tong, H. (1977) Some comments on the Canadian lynx data.
Journal of the Royal Statistical Society A 140, 432-436.
}
\seealso{
\code{\link[=SelectModel]{SelectModel()}}, \code{\link[=GetFitARpLS]{GetFitARpLS()}}, \code{\link[=leaps]{leaps()}}.
}
\author{
A.I. McLeod.
}
