% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRho.R
\name{getRho}
\alias{getRho}
\title{Normalized rho unit root test statistic}
\usage{
getRho(ans)
}
\arguments{
\item{ans}{output from FitAR.}
}
\value{
Value of the test statistic.
}
\description{
Utility function used by UnitRootTest.
}
\examples{
z <- cumsum(rnorm(100))
ans <- FitAR(z, p=1)
getRho(ans)

}
\seealso{
\code{\link[=getT]{getT()}}, \code{\link[=UnitRootTest]{UnitRootTest()}}.
}
\author{
A.I. McLeod.
}
