% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getT.R
\name{getT}
\alias{getT}
\title{t-statistic for unit root test}
\usage{
getT(ans)
}
\arguments{
\item{ans}{output from FitAR.}
}
\value{
Value of the test statistic.
}
\description{
Utility function used by UnitRootTest.
}
\examples{
z <- cumsum(rnorm(100))
ans <- FitAR(z, p=1)
getT(ans)

}
\seealso{
\code{\link[=getRho]{getRho()}}, \code{\link[=UnitRootTest]{UnitRootTest()}}.
}
\author{
A.I. McLeod.
}
