% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glog.R
\name{glog}
\alias{glog}
\title{glog transformation}
\usage{
glog(x, a = 1, InverseQ = FALSE)
}
\arguments{
\item{x}{numeric vector of data.}

\item{a}{additive constant, often 1.}

\item{InverseQ}{inverse glog.}
}
\value{
transformated data.
}
\description{
The glog is a better behaved log transformation when some data values are
zero or just near zero.
}
\details{
Basic properties of the glog transformation are illustrated in
the Mathematica notebook glog.nb and its pdf version glog.pdf which are
available in the package directory doc.
}
\examples{
# usual log transformation doesn't work
all(is.finite(log(sunspot.month)))
# either shifted log
all(is.finite(log(sunspot.month+1)))
# or glog works
all(is.finite(glog(sunspot.month)))
# but glog may be better, especially for values <1 but >=0

}
\references{
W. Huber, A. von Heydebreck, H. Sultmann, A. Poustka, and
M. Vingron. Variance stablization applied to microarray data calibration
and to quantification of differential expression. Bioinformatics, 18: S96-S10 2002.
}
\seealso{
\code{\link[=bxcx]{bxcx()}}.
}
\author{
A.I. McLeod.
}
