% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InformationMatrixAR.R
\name{InformationMatrixAR}
\alias{InformationMatrixAR}
\title{Information Matrix for AR(p)}
\usage{
InformationMatrixAR(phi)
}
\arguments{
\item{phi}{vector of length p corresponding to the AR(p) coefficients.}
}
\value{
a p-by-p Toeplitz matrix, p = length(phi).
}
\description{
The Fisher large-sample information matrix per observation for the
p coefficients in an AR(p) is computed.
}
\details{
The Fisher information matrix is computed as the covariance matrix of an AR(p)
process with coefficients given in the argument \code{phi} and with unit innovation
variance. The \code{TacvfAR} function is used to compute the necessary autocovariances.
\code{FitAR} uses \code{InformationMatrixAR} to obtain estimates of the standard
errors for the estimated parameters in the case of the full AR(p) model.
}
\examples{
InformationMatrixAR(c(1.8,-0.6))

}
\references{
McLeod, A.I. and Zhang, Y. (2006). Partial autocorrelation
parameterization for subset autoregression. Journal of Time Series Analysis, 27, 599-612.
}
\seealso{
\code{\link[=FitAR]{FitAR()}}, \code{\link[=InformationMatrixARp]{InformationMatrixARp()}}, \code{\link[=TacvfAR]{TacvfAR()}}, \code{\link[=InformationMatrixARz]{InformationMatrixARz()}}.
}
\author{
A.I. McLeod and Y. Zhang.
}
