% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InformationMatrixARz.R
\name{InformationMatrixARz}
\alias{InformationMatrixARz}
\title{Fisher Information Matrix Subset Case, ARz}
\usage{
InformationMatrixARz(zeta, lags)
}
\arguments{
\item{zeta}{vector of coefficients, ie. partial autocorrelations at lags
specified in the argument \code{lags}.}

\item{lags}{lags in subset model, same length as zeta argument.}
}
\value{
a p-by-p Toeplitz matrix, p = length(zeta).
}
\description{
Computes the large-sample Fisher information matrix per observation for
the AR coefficients in a subset AR when parameterized by the partial autocorrelations.
}
\details{
The details of the computation are given in McLeod and Zhang (2006, eqn 13).
\code{FitAR} uses \code{InformationMatrixARz} to obtain estimates of the standard
errors of the estimated parameters in the subset AR model when partial
autocorrelation parameterization is used.
}
\examples{
# Information matrix for ARz(1,4) with parameters 0.9 and 0.9.
InformationMatrixARz(c(0.9, 0.9), lags=c(1,4))

}
\references{
McLeod, A.I. and Zhang, Y. (2006). Partial autocorrelation
parameterization for subset autoregression. Journal of Time Series Analysis, 27, 599-612.
}
\seealso{
\code{\link[=FitAR]{FitAR()}}, \code{\link[=InformationMatrixAR]{InformationMatrixAR()}}, \code{\link[=InformationMatrixARp]{InformationMatrixARp()}}.
}
\author{
A.I. McLeod and Y. Zhang.
}
