% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JacobianK.R
\name{JacobianK}
\alias{JacobianK}
\title{Internal Utility Function}
\usage{
JacobianK(zeta, k)
}
\arguments{
\item{zeta}{partial autocorrelations.}

\item{k}{k-th Jacobian.}
}
\value{
Matrix.
}
\description{
The matrix defined in eqn. (10) of McLeod and Zhang (2006).
Used by the function Jacobian.
}
\examples{
JacobianK(rep(0.8,4),3)

}
\references{
McLeod, A.I. and Zhang, Y. (2006). Partial autocorrelation
parameterization for subset autoregression. Journal of Time Series Analysis, 27, 599-612.
}
\seealso{
\code{\link[=Jacobian]{Jacobian()}}.
}
\author{
A.I. McLeod.
}
