% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JarqueBeraTest.R
\name{JarqueBeraTest}
\alias{JarqueBeraTest}
\title{Jarque-Bera Normality Test}
\usage{
JarqueBeraTest(z)
}
\arguments{
\item{z}{vector of data.}
}
\value{
The returned values are as follow:
\itemize{
\item \code{LM} value of the LM statistic.
\item \code{pvalue} p-value.
}
}
\description{
A powerful omnibus test for normality.
}
\details{
This test is derived as a Lagrange multiplier test for normal
distribution in the family of Pearson distributions (Jarque and Bera, 1987).
}
\examples{
# some normal data
z<-rnorm(100)
JarqueBeraTest(z)
# some skewed data
z<-rexp(100)
JarqueBeraTest(z)
# some thick tailed data
z<-rt(100,5)
JarqueBeraTest(z)

}
\references{
Jarque, C.M. and Bera, A.K. (1987). A Test for Normality of
Observations and Regression Residuals. International Statistical Review 55, 163-172.
}
\author{
A.I. McLeod.
}
