% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PacfPlot.R
\name{PacfPlot}
\alias{PacfPlot}
\title{Plot Partial Autocorrelations and Limits}
\usage{
PacfPlot(z, lag.max = 15, ...)
}
\arguments{
\item{z}{time series.}

\item{lag.max}{maximum lag, P.}

\item{...}{optional parameters passed through to plot.}
}
\value{
No value is returned. Graphical output is produced as side-effect.
}
\description{
The sample partial autocorrelations and their individual 95 percent confidence
intervals are plotted under the assumption the model is contained in an AR(P),
where P is a specified maximum order.
}
\details{
The Burg algorithm is used to estimate the PACF.
}
\examples{
# For the log(lynx) series and taking lag.max=15, the PacfPlot and
# the minimum BIC subset selection produce the same result.
z<-log(lynx)
PacfPlot(z)
SelectModel(z,lag.max=15,ARModel="ARz",Best=1,Criterion="BIC")

}
\references{
McLeod, A.I. and Zhang, Y. (2006). Partial autocorrelation
parameterization for subset autoregression. Journal of Time Series Analysis, 27, 599-612.
}
\seealso{
\code{\link[=ar.burg]{ar.burg()}}, \code{\link[=pacf]{pacf()}}.
}
\author{
A.I. McLeod and Y. Zhang.
}
