% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PacfToAR.R
\name{PacfToAR}
\alias{PacfToAR}
\title{Transform from PACF Parameters to AR Coefficients}
\usage{
PacfToAR(zeta)
}
\arguments{
\item{zeta}{vector of AR PACF parameters.}
}
\value{
Vector of AR coefficients.
}
\description{
Transforms AR partical autocorrelation function (PACF) parameters to AR
coefficients based on the Durbin-Levinson recursion.
}
\details{
See Mcleod and Zhang (2006).
}
\examples{
somePACF<-c(0.5,0.6,0.7,0.8,-0.9,-0.8)
someAR<-PacfToAR(somePACF)
test<-ARToPacf(someAR)
# this should be very small
sum(abs(test-somePACF))

}
\references{
McLeod, A.I. and Zhang, Y. (2006). Partial autocorrelation
parameterization for subset autoregression. Journal of Time Series Analysis, 27, 599-612.
}
\seealso{
\code{\link[=InvertibleQ]{InvertibleQ()}}, \code{\link[=PacfToAR]{PacfToAR()}}.
}
\author{
A.I. McLeod and Y. Zhang.
}
