% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.FitAR.R
\name{plot.FitAR}
\alias{plot.FitAR}
\title{Plot Method for "FitAR" Object}
\usage{
plot(x, terse=TRUE, ...) ## S3 method for class 'FitAR'
}
\arguments{
\item{x}{object of class "FitAR".}

\item{terse}{if TRUE, only one graph is produced, otherwise many diagnostic plots.}

\item{...}{optional arguments.}
}
\value{
No value is returned. Plots are produced as side-effect.
}
\description{
Diagnostic plots: portmanteau p-values; residual autocorrelation plot; normal
probability plot and Jarque-Bera test; spectral density function.
}
\note{
When terse=FALSE, numerous graphs are produced.
Turn on recording to be able to page back and forth between them.
}
\examples{
obj<-FitAR(SeriesA, c(1,2,6,7))
plot(obj)

}
\references{
McLeod, A.I. and Zhang, Y. (2006). Partial autocorrelation
parameterization for subset autoregression. Journal of Time Series Analysis, 27, 599-612.
}
\seealso{
\code{\link[=summary.FitAR]{summary.FitAR()}}, \code{\link[=FitAR]{FitAR()}}, \code{\link[=JarqueBeraTest]{JarqueBeraTest()}}, \code{\link[=RacfPlot]{RacfPlot()}}, \code{\link[=LBQPlot]{LBQPlot()}}.
}
\author{
A.I. McLeod and Y. Zhang.
}
