% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Selectmodel.R
\name{plot.Selectmodel}
\alias{plot.Selectmodel}
\title{Subset AR Graph for "Selectmodel" Object}
\usage{
plot(x, ...) ## S3 method for class 'Selectmodel'
}
\arguments{
\item{x}{out from \code{SelectModel}.}

\item{...}{optional arguments.}
}
\value{
No value. Plot produced as side-effect.
}
\description{
A graphical depiction is given of the output from \code{SelectModel}.
}
\details{
The relative plausibility of Model A vs the best Model B, is defined as
\eqn{R=e^{(AIC_B-AIC_A)/2}}. Values of R less than 1 R is defined similarly
if the BIC/UBIC criterion is used.
}
\examples{
# takes about 10 seconds
## Not run: 
out<-SelectModel(log(Willamette),lag.max=150,ARModel="AR",Best=5,Criterion="AIC")
plot(out)
## End(Not run)

}
\seealso{
\code{\link[=SelectModel]{SelectModel()}}.
}
\author{
A.I. McLeod.
}
