% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.FitAR.R
\name{print.FitAR}
\alias{print.FitAR}
\title{Print Method for "FitAR" Object}
\usage{
print(x, ...) ## S3 method for class 'FitAR'
}
\arguments{
\item{x}{object of class "FitAR".}

\item{...}{optional arguments.}
}
\value{
A terse summary is displayed.
}
\description{
A terse summary is given.
}
\examples{
data(SeriesA)
FitAR(SeriesA, c(1,2,6,7))

}
\references{
McLeod, A.I. and Zhang, Y. (2006). Partial autocorrelation
parameterization for subset autoregression. Journal of Time Series Analysis, 27, 599-612.
}
\seealso{
\code{\link[=summary.FitAR]{summary.FitAR()}}.
}
\author{
A.I. McLeod.
}
