% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.FitAR.R
\name{residuals.FitAR}
\alias{residuals.FitAR}
\title{Extract Residuals from "FitAR" Object}
\usage{
residuals(object, ...) ## S3 method for class 'FitAR'
}
\arguments{
\item{object}{object of class "FitAR".}

\item{...}{optional arguments.}
}
\value{
Vector of residuals.
}
\description{
Method function.
}
\examples{
out<-FitAR(SeriesA, c(1,2,6,7))
resid(out)

}
\seealso{
\code{\link[=FitAR]{FitAR()}}.
}
\author{
A.I. McLeod.
}
