% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdfplot.ar.R
\name{sdfplot.ar}
\alias{sdfplot.ar}
\title{Autoregressive Spectral Density Estimation for "ar"}
\usage{
sdfplot(obj, ...) ## S3 method for class 'ar'
}
\arguments{
\item{obj}{class "ar" object, output from \code{ar}.}

\item{...}{optional arguments.}
}
\value{
Plot is produced using plot. Matrix with 2 columns containing
the frequencies and spectral density is returned invisibly.
}
\description{
Method for class "ar" for sdfplot.
}
\examples{
# Fit AR(p) using AIC model selection and Burg estimates.
# Plot spectral density estimate.
ans<-ar(lynx, lag.max=20)
sdfplot(ans)

}
\seealso{
\code{\link[=ARSdf]{ARSdf()}}, \code{\link[=sdfplot]{sdfplot()}}, \code{\link[=sdfplot.FitAR]{sdfplot.FitAR()}}, \code{\link[=sdfplot.ts]{sdfplot.ts()}}.
}
\author{
A.I. McLeod.
}
