% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdfplot.Arima.R
\name{sdfplot.Arima}
\alias{sdfplot.Arima}
\title{Spectral Density of Fitted ARIMA Model}
\usage{
sdfplot(obj, ...) ## S3 method for class 'Arima'
}
\arguments{
\item{obj}{object class "Arima".}

\item{...}{optional arguments.}
}
\value{
Plot is produced using plot. Matrix with 2 columns containing
the frequencies and spectral density is returned invisibly.
}
\description{
Method for class "Arima" for sdfplot.
}
\examples{
sdfplot(SeriesA,c(1,0,1))

}
\seealso{
\code{\link[=ARSdf]{ARSdf()}}, \code{\link[=sdfplot]{sdfplot()}}, \code{\link[=sdfplot.FitAR]{sdfplot.FitAR()}}, \code{\link[=sdfplot.ts]{sdfplot.ts()}}.
}
\author{
A.I. McLeod.
}
