% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdfplot.FitAR.R
\name{sdfplot.FitAR}
\alias{sdfplot.FitAR}
\title{Autoregressive Spectral Density Estimation for "FitAR"}
\usage{
sdfplot(obj, ...) ## S3 method for class 'FitAR'
}
\arguments{
\item{obj}{object class "FitAR".}

\item{...}{optional arguments.}
}
\value{
Plot is produced using plot. Matrix with 2 columns containing
the frequencies and spectral density is returned invisibly.
}
\description{
Methods function for sdfplot. Autoregressive spectral density function
estimation using the result output from FitAR.
}
\examples{
# Use AIC to select best subset model to fit to lynx data and
# plot spectral density function
pvec<-SelectModel(SeriesA, ARModel="ARp", lag.max=10, Best=1)
ans<-FitAR(SeriesA, pvec)
sdfplot(ans)

# plot sdf and put your own title
z<-c(SeriesA)
pvec<-SelectModel(z, ARModel="ARp", lag.max=10, Best=1)
ans<-FitAR(z, pvec)
sdfplot(ans)
title(main="Example SDF")

}
\seealso{
\code{\link[=sdfplot]{sdfplot()}}, \code{\link[=FitAR]{FitAR()}}.
}
\author{
A.I. McLeod.
}
