% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdfplot.numeric.R
\name{sdfplot.numeric}
\alias{sdfplot.numeric}
\title{Autoregressive Spectral Density Estimation for "numeric"}
\usage{
sdfplot(obj, ...) ## S3 method for class 'numeric'
}
\arguments{
\item{obj}{object, class"numeric", a vector.}

\item{...}{optional arguments.}
}
\value{
Plot is produced using plot. Matrix with 2 columns containing
the frequencies and spectral density is returned invisibly.
}
\description{
Method function for vectors, class "numeric".
}
\examples{
sdfplot(lynx)

}
\seealso{
\code{\link[=sdfplot]{sdfplot()}}.
}
\author{
A.I. McLeod.
}
