% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdfplot.R
\name{sdfplot}
\alias{sdfplot}
\title{Autoregressive Spectral Density Estimation}
\usage{
sdfplot(obj, ...)
}
\arguments{
\item{obj}{input object.}

\item{...}{optional arguments.}
}
\value{
Plot is produced using plot. Matrix with 2 columns containing
the frequencies and spectral density is returned invisibly.
}
\description{
Generic function. Methods are available for "FitAR", "ar", "Arima", "ts" and "numeric".
}
\examples{
# Example 1
# Use AIC to select best subset model to fit to lynx data and
# plot spectral density function
pvec<-SelectModel(SeriesA, ARModel="ARp", lag.max=10, Best=1)
ans<-FitAR(SeriesA, pvec)
sdfplot(ans)

# Example 2
# Fit ARMA and plot sdf
ans<-arima(SeriesA, c(1,0,1))
sdfplot(ans)

# Example 3
# Fit ARz model using AIC to monthly sunspots and plot spectral density
# Warning: this may take 10 minutes or so.
## Not run: 
pvec<-SelectModel(sunspots, lag.max=200, ARModel="ARz", Criterion="AIC", Best=1)
ans<-FitAR(sunspots, pvec)
sdfplot(ans)
## End(Not run)#' 

}
\seealso{
\code{\link[=sdfplot]{sdfplot()}}, \code{\link[=FitAR]{FitAR()}}.
}
\author{
A.I. McLeod.
}
