% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateGaussianAR.R
\name{SimulateGaussianAR}
\alias{SimulateGaussianAR}
\title{Autoregression Simulation}
\usage{
SimulateGaussianAR(phi, n = 100, InnovationVariance = 1)
}
\arguments{
\item{phi}{vector containing AR coefficients.}

\item{n}{length of time series.}

\item{InnovationVariance}{innovation variance.}
}
\value{
A vector of length n, the simulated series.
}
\description{
Simulate a mean-zero stationary Gaussian AR(p) time series.
}
\details{
The p initial values are simulated using the appropriate multivariate
distribution as was suggested in McLeod (1975). The R function \code{rnorm()} is used.
}
\note{
If the process is non-stationary, then random initial values are used
determined by the first p innovations.
}
\examples{
# Percival and Walden (1993, p.46) illustrated a time series with a
# very peaked spectrum with the AR(4) with coefficients
# c(2.7607,-3.8106,2.6535,-0.9238) with NID(0,1) innovations.
z<-SimulateGaussianAR(c(2.7607,-3.8106,2.6535,-0.9238),1000)
library(lattice)
TimeSeriesPlot(z)

}
\references{
McLeod, A.I. (1975), Derivation of the theoretical autocorrelation function of
autoregressive moving-average time series, Applied Statistics 24, 255–256.

Percival, D.B. and Walden, A.T. (1993), Spectral Analysis for Physical Applications.
}
\seealso{
\code{\link[=Boot.FitAR]{Boot.FitAR()}}.
}
\author{
A.I. McLeod.
}
