% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.FitAR.R
\name{summary.FitAR}
\alias{summary.FitAR}
\title{Summary Method for "FitAR" Object}
\usage{
summary(object, ...) ## S3 method for class 'FitAR'
}
\arguments{
\item{object}{"FitAR" object.}

\item{...}{optional arguments.}
}
\value{
A printed summary is given.
}
\description{
summary for "FitAR" object.
}
\examples{
data(SeriesA)
out<-FitAR(SeriesA, c(1,2,6,7))
summary(out)

}
\seealso{
\code{\link[=print.FitAR]{print.FitAR()}}, \code{\link[=FitAR]{FitAR()}}.
}
\author{
A.I. McLeod.
}
