% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TacvfMA.R
\name{TacvfMA}
\alias{TacvfMA}
\title{Theoretical Autocovariances for Moving Average Process}
\usage{
TacvfMA(theta, lag.max = 20)
}
\arguments{
\item{theta}{q parameters in MA(q).}

\item{lag.max}{number of lags required.}
}
\value{
Vector of length q+1 containing the autocovariances at lags 0,1,...,lag.max.
}
\description{
The theoretical autocovariance function of a MA(q) with unit variance is computed.
}
\details{
The first q+1 values are determined using a matrix multiplication -
avoiding a loop. The remaining values set to zero.
}
\note{
See Details in \code{TacvfAR} for why we prefer to use this algorithm instead of \code{ARMAacf}.
}
\examples{
TacvfMA(c(1.8,-0.9), 10)

}
\references{
McLeod, A.I. and Zhang, Y. (2006), Partial autocorrelation parameterization
for subset autoregression. Journal of Time Series Analysis, 27, 599-612.
}
\seealso{
\code{\link[=ARMAacf]{ARMAacf()}}, \code{\link[=TacvfAR]{TacvfAR()}}.
}
\author{
A.I. McLeod and Y. Zhang.
}
