% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarianceRacfAR.R
\name{VarianceRacfAR}
\alias{VarianceRacfAR}
\title{Covariance Matrix Residual Autocorrelations for AR}
\usage{
VarianceRacfAR(phi, MaxLag, n)
}
\arguments{
\item{phi}{vector of AR coefficients.}

\item{MaxLag}{covariance matrix for residual autocorrelations at
lags 1 ,..., m, where m = MaxLag is computes.}

\item{n}{length of time series.}
}
\value{
The m-by-m covariance matrix of residual autocorrelations at
lags 1, ..., m, where m = MaxLag.
}
\description{
Computes the variance-covariance matrix for the residual autocorrelations in an AR(p).
}
\details{
The covariance matrix for the residual autocorrelations is derived in McLeod
(1978, eqn. 15) for the general ARMA case. With this function one can obtain
the standard deviations of the residual autocorrelations which can be used for
diagnostic checking with \code{RacfPlot}.
}
\note{
The derivation assumes normality of the innovations, mle estimation of the
parameters and a known mean-zero time series. It is easily seen that
the same result still holds for IID innovations with mean zero and finite
variance, any first-order efficient estimates of the parameters including
the AR coefficients and mean.
}
\examples{
VarianceRacfAR(0.5,5,100)

}
\references{
McLeod, A.I. (1978), On the distribution and applications of residual
autocorrelations in Box-Jenkins modelling, Journal of the Royal
Statistical Society B, 40, 296–302.
}
\seealso{
\code{\link[=VarianceRacfARp]{VarianceRacfARp()}}, \code{\link[=VarianceRacfARz]{VarianceRacfARz()}}, \code{\link[=RacfPlot]{RacfPlot()}}.
}
\author{
A.I. McLeod.
}
